moodmagic/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.12.0
requests==2.31.0
stripe==7.14.0
firebase-admin==6.4.0
python-jose[cryptography]==3.3.0 